Fix select statements in identity/sgx/BUILD
This changes the select statements to include the SGX implementation if
building with the enclave toolchain, and the fake implementation if
building with the non-enclave toolchain. Previously, these selects were
based on config setting, which was resulted in some false positives for
the SGX select (e.g., the SGX select statement was triggered if the
library was used in the deps of an enclave_test rule).

PiperOrigin-RevId: 253338704
Change-Id: I095a31d489700aa4d99e6d44358e87626532c171