Add signal handling enclave call in primitives/
The signal handling enclave call can not be implemented via the
generic primitives Client::EnclaveCall for SGX given that the
edger8r ecall declaration needs to be prepended with
`exception_handler`. Client::DeliverSignal is added to the primitives
interface under the assumption of future non-SGX backends also
requiring a specialized enclave call for implementing signal handling.

PiperOrigin-RevId: 258425373
Change-Id: Iafbed6ff6e4ebcc0999b678c790ad4e0e859ec7d