Move PCE constants into separate cc_library
This change moves the PCE constants out of sgx_pce.h and into a separate
header, sgx_pce_constants.h, that can be compiled into a standalone
cc_library.

The motivation behind this change is that PCE constants need to be used
by enclaves that are generating REPORTS for the PCE protocols. However,
these constants are currently defined in the same header as the PCE
interface itself, which can only be invoked from outside an enclave.

PiperOrigin-RevId: 258867158
Change-Id: Ia01186b81577de7adf682f3f0f9f0fe23a53c11a