Update pce_util to depend on pce_constants lib
The pce_util library needs to be invoked inside an enclave so it cannot
depend on pce_wrapper, which relies on URTS. pce_util only needs the
definitions of PCE constants defined in the pce_constants library.

Also add a test for this library that runs inside an enclave. The
addition of std::move on a couple lines is to account for toolchain
differences.

PiperOrigin-RevId: 258880090
Change-Id: Ifcecd235bafad65dd6dee0b6b3c9e7d49ecffb29