Modify type conversion function for syscalls
This change introduces a parameter 'or_input_to_default_value' to
modify the generated type conversions function. The parameter allows
input value to the conversion function to be OR'd with the default
value on host or newlib in case no matching enum value is found.

The change also defines constants for errnos to generate the type
conversion functions for the same.

PiperOrigin-RevId: 261185434
Change-Id: I4cb53556697a20b617b432964d9e8177075a25c3