Add KSS-related attributes in SGX enclave config
This change updates the sgx_enclave_configuration rule to allow setting
whether or not KSS is enabled for the enclave, as well as the values of
ISVFAMILYID and ISVEXTPRODID.

There are various restrictions around how and when these attributes can
be configured. Most notably, ISVFAMILYID and ISVEXTPRODID cannot be set
unless KSS is also enabled. Additionally, the sgx_sign_tool required
that the ISVFAMILYID and ISVEXTPRODID are formatted as hex-encoded
16-byte strings. This change introduces validation logic to ensure that
bad configurations are not created.

Note that running an enclave with KSS-related features enabled on a
system that does not support KSS will cause EINIT-time failures.

PiperOrigin-RevId: 263214275
Change-Id: Ic6a995f4845a0e5bcfe559bedf76d4f89605f45b