Add changes to newlib's math.h for experiment
The missing long double math functions in newlib cause problems with
the LLVM project's libc++. This commit does not integrate MUSL as a
substitute libc yet, but it does change the header for some
experimentation we're doing in this area. Clang and libc++ support will
come via multiple incremental changes like this one.

PiperOrigin-RevId: 263244708
Change-Id: Id6362b2dd4c7f1e7ac7a1cdd7d9d7a0702cdd8dc