Lazily initialize host_call library
This removes init_host_calls(), which is required prior to
using the host call library. Also removes the need to initialize
host calls in enc_untrusted_nanosleep() as a special case.

Introduces an intermediary wrapper,
ensure_initialized_and_dispatch_syscall() which initializes the
system_call library if needed and forwards the arguments to
enc_untrusted_syscall.

PiperOrigin-RevId: 263608152
Change-Id: Ifdca037e73abfda722ee62e99b99b980753e69e1