Migrate enc_untrusted_sendmsg
Since sendmsg refers to a struct msghdr, which includes a scattered
array, this method is implemented as a special case rather than using
the existing enc_untrusted_syscall implementation. This is implemented
by first combine the scattered array into a single buffer on the trusted
side, then pass it along with other params to the untrusted side. The
corresponding exit handler creates a new msghdr struct on the host, make
it point to the buffers, and set the array size to 1, then call sendmsg
on the host with this msghdr.

PiperOrigin-RevId: 264201806
Change-Id: I6021f3fc94d9a3b4052b3cda0f2dc45bb1bae5df