Introduce RegisterExitHandlers for primitives
This introduces a virtual method RegisterExitHandlers() which may
be used for registering exit handlers for the primitives layer,
which are specific to a particular backend.
The change also provides an untrusted handler for SGX backend which
is used for thread creation, and registers it using the overridden
method  RegisterExitHandlers() for SGX. We need no such method for
sim backend, since thread creation can occur without exiting the
enclave.

PiperOrigin-RevId: 264230039
Change-Id: Id3f2687aabff781baa99480b37268be8f6d0ac24