Migrate thread creation to primitives layer
This change removes enc_untrusted_create_thread() and its
associated ocall and delegates the call to a CreateThread()
function in primitives layer.
Thread creation might have different implementation for various
backends. For sim, the enclave need not be exited. For SGX,
CreateThread() performs an UntrustedCall to a CreateThread
handler which makes an EnclaveCall to DonateThread with the
created thread to register itself inside the enclave and start
the thread.

Also with this change, we get rid of the dependency on
EnclaveManager on the untrusted side to get the Client to make
the EnclaveCall. We now instead use the primitive client provided
directly as an argument to the untrusted CreateThread handler to
perform the EnclaveCall.

PiperOrigin-RevId: 264448076
Change-Id: If297c6deb327ad2246c4195769ea644de91551f2