Remove peer_identity_util
This change removes the peer_identity_util library. This library
provided ACL-evaluation logic using an EnclaveAuthContext, but this
logic is now built into EnclaveAuthContext. As a result, this
functionality is no longer needed here.

This library also provided a utility to extract an SGX CodeIdentity.
The one usage of this function was replaced with a local helper
function. Furthermore, this function was not independently tested so it
does not make sense to expose it from a util library.

PiperOrigin-RevId: 264690722
Change-Id: I493c21903900b296bbce91946b827e491a738ac5