Implement SgxIdentity code_identity_util methods
This commit implements `SgxIdentity`-compatible versions of existing
methods in `code_identity_util` that operate on `CodeIdentity`. The
majority of these implementations simply fill out the required
`SgxMachineConfiguration` information, then call out to the existing
utility methods to fill in the `CodeIdentity` components.

The plan moving forward is to keep the "legacy" methods around until
dependencies on `code_identity_util` have all been migrated to pass
`SgxIdentity` messages rather than `CodeIdentity` messages; after this,
we will be able to move the full implementations into the new methods
and deprecate the older ones.

This commit also introduces the `kSgxIdentityVersionString` constant,
which is used in the parsing/deserialization methods, in order to
maintain backwards-compatibility with existing serialized `identity`
that are `CodeIdentity`s rather than `SgxIdentity`s.

PiperOrigin-RevId: 265142130
Change-Id: Iacebfe80213f7b2252fcf2f35fe86b7a76e16517