Use SgxIdentity representation in SelfIdentity
This commit replaces the `identity` field of `SelfIdentity` (which is a
CodeIdentity) with a new `sgx_identity` field, as part of the ongoing
refactoring of all `CodeIdentity`s into `SgxIdentity`s. To accomplish
this, all callers of `SelfIdentity` were also updated to operate on the
new `sgx_identity` field instead.

PiperOrigin-RevId: 265151246
Change-Id: Iaf9eeabecbd59656bb9f7a1d80d1cc59ca1fbb26