Reduce randomized code_identity_util_test runs
Currently, the parsing tests for randomized identities / match specs /
expectations run a seemingly arbitrary 10000 times each. This leads to
a fairly long-running test and an unexplained magic number.

Noting that the generation of valid/invalid identities are the bounding
factor, we can solve the "coupon collector's problem" for n=48 to land
on a different magic number of random trials that will cover parsing all
of the possible variants that we care about (specifically 500, which
reduces the runtime of these tests by an order of magnitude).

PiperOrigin-RevId: 266254236
Change-Id: I8e1b05e6fd5a7d535957724a8ddba3a5a8558864