Add offline signing features for sgx enclaves
This change deprecates the sgx_enclave rule.  The sgx_enclave rule is
not clear from its name that the resulting binary is signed by a debug
key. Furthermore there was no easy way to sign an enclave in an offline
manner without directly invoking the Intel signing tool.

This change adds two rules to make offline signing easier to work with
Bazel:

1. sgx_generate_sigstruct produces the binary material
from a config and unsigned enclave that must be signed.

2. sgx_signed_enclave combines the output of the previous rule with a
signature file, a public key file to verify the signature, and produces
a signed enclave binary.

This change additionally exposes the two rules that make up the
deprecated sgx_enclave rule:

1. sgx_unsigned_enclave defines the cc_binary that should be built
with the Asylo toolchain.

2. sgx_debug_enclave uses a debug key in the source tree to sign an
unsigned enclave binary.

The added steps are updated in the sgx_release_enclaves documentation.

PiperOrigin-RevId: 266397838
Change-Id: Icd32db578b532fe382389e2bc7ca943a033af055