Add legacy support to code_identity_util methods
This commit adds an `is_legacy` parameter to the methods in
`code_identity_util` that consume `SgxIdentity(MatchSpec|Expectation)`,
since we currently lack a way of determining whether a match spec
belonged to a "legacy" CodeIdentity-based identity or was created
after the migration to SgxIdentity.

We are okay with SgxMachineConfiguration being absent from any legacy
match specs, but want that field to be present going forward, which is
what differentiates the two code paths.

There is no need to add an `is_legacy` parameter to the validity checks
for SgxIdentity, because this method only verifies the value of CPU SVN
iff that field is present in the message (a "legacy" identity will still
be verified properly since the CPU SVN will simply be absent).

PiperOrigin-RevId: 267011285
Change-Id: I24dc9b9d161418bdad8b317f7fb7128aad0aef31