Change clock_gettime to make an exit call
Currently clock_gettime is implemented by shared resources, based on the
assumption of access to untrusted memory. This needs to be changed for
primitives, as the assumption is not true for remote backend, or
possibly other backends.

Also move the check of CLOCK_MONOTONIC to ensure checking it's not going
backwards in the new implementation.

PiperOrigin-RevId: 268339262
Change-Id: Ifcd607bc48d1607c5eb310b2a0dbc6fcd6fba2ef