Add MatchAndExplain method to IdentityExpMatcher
This change adds a new MatchAndExplain method to the
IdentityExpectationMatcher interface. This method is identical to the
Match method, but with an additional |explanation| parameter that is
populated with an explanation of why a match returned false.

To avoid breaking existing subclasses, IdentityExpectationMatcher
provides a default implementation of MatchAndExplain that just calls
Match, ignoring the |explanation| parameter.

PiperOrigin-RevId: 268728865
Change-Id: Iace630e03a8df59102b03fc48fb99c2564bf9752