Migrate SecretSealer to use SgxIdentity
This commit updates the implementations of `local_secret_sealer`
methods to operate on SgxIdentity instead of CodeIdentity. To maintain
backwards compatibility with secrets sealed in the CodeIdentity version
of the codebase, we determine whether we're dealing with a "legacy
secret" by checking the `client_acl` of the header in ParseKey-
GenerationParamsFromSealedSecretHeader. If the `version` field of the
reference identity is not set, we choose the legacy code path.

Another change that this commit makes is deprecating use of the
`additional_info` field in SealingRootInformation. Because CPUSVN is
now contained in SgxIdentity, moving `cipher_suite` into its own field
in SealingRootInformation removes the need for `additional_info` at
all. This proto change is also a chance to use asylo::AeadScheme
instead of sgx::CipherSuite.

Finally, this commit removes `cpusvn` as an output parameter from
methods in `local_secret_sealer_helpers`, since the new implementation
populates the CPUSVN value within SgxIdentityExpectation instead.

PiperOrigin-RevId: 268933639
Change-Id: I5d8406e8ffb64cff49ee72bbd80b03e84d1ae593