Fix MatchIdentityToExpectation matching logic
This commit fixes a bug with the logic of MatchIdentityToExpectation.
The function was incorrectly checking for the presence of various
SgxMachineConfigurationMatchSpec fields being present to determine
whether those values in the identity should match the expectation's
reference identity, when in reality it should have been checking that
the fields in the match spec were set to true (and not just present).

The effect of this was that the matching was being stricter than
necessary, potentially rejecting identities that should actually match
in the event that the match spec fields had been set to false.

PiperOrigin-RevId: 269666551
Change-Id: Ife472e06108744f073712d6cee127e95b537f915