Fix SgxIdentityExpectationMatcher legacy logic
This commit fixes a bug with the way in which the MatchAndExplain method
in SgxIdentityExpectationMatcher deal with legacy expectations. Because
it should be possible to have a legacy identity with a non-legacy
expectation, the `is_legacy` flag should be determined based on the
presence of the `version` field in the reference identity of the
expectation rather than the identity itself.

PiperOrigin-RevId: 269697866
Change-Id: If949693202d37e8b62cd540f9e5fce9c4c1a337f