Use SgxIdentity in RemoteAssertionGeneratorEnclave
This commit updates the GetEnclaveIdentityOutput message in
RemoteAssertionGeneratorEnclave to use SgxIdentity directly rather than
a combination of CodeIdentity and CpuSvn. This change was enabled by
the completed implementation of SetSelfSgxIdentity, which now correctly
sets the CPUSVN value from the SelfIdentity.

In addition, changes were also made to SgxInfrastructuralEnclaveManager
to accommodate the updated output message format.

PiperOrigin-RevId: 270128863
Change-Id: I752fe74015acfcb9f5cd2bd904308f37f98ec07d