Implement Is{Inside,Outside}Enclave
Previously, we defined IsTrustedExtent which returned true if every byte
of a memory range is inside the enclave. However !IsTrustedExtent is in
general not sufficient to ensure that a buffer is untrusted, since "Not
every byte inside enclave" is true for "Some bytes inside enclave."

This patch also changes the terminology to "IsInsideEnclave"
and "IsOutsideEnclave." This is intended to communicate the intent and
semantics more clearly.

Thanks to David Oswald <d.f.oswald@cs.bham.ac.uk> and Jo Van Bulck
<Jo.VanBulck@cs.kuleuven.be> for uncovering and reporting this issue.

PiperOrigin-RevId: 270739874
Change-Id: I84baa5e1ce865d4596ab8177946741a78b1f8059