Automatically calculate some DCAP output sizes
Previously, some output buffers to the dcap wrapper had to be set to
some minimum length. There is no reason to require this, as the library
can determine the appropriate buffer output size based on use case and
allocate memory on its own.

This means that the encrypted PPID and signed report outputs from the
PCE will now be calculated on behalf of the user.

This has the bonus of optimizing away some buffer reallocs.

PiperOrigin-RevId: 270918772
Change-Id: I6deff9f580c7ced929b693aafe3417acc899139f