Remove unused CodeIdentity utility methods
This commit takes an initial pass at either removing or making internal
some of the methods in `code_identity_util` that consume a CodeIdentity,
as part of the ongoing migration from CodeIdentity to SgxIdentity.

The changes in this commit are the "low-hanging fruit changes" that do
not require complex changes to the rest of Asylo; roughly speaking, any
methods that are not required by any current unit tests or other parts
of the library were either moved to the anonymous namespace within
`code_identity_util` if called by the newer SgxIdentity methods, or
deleted altogether if obsolete.

The tests that were removed in `code_identity_util_test` were ones whose
SgxIdentity counterparts sufficiently tested the functionality, and so
they were safe to remove without affecting overall code coverage.

PiperOrigin-RevId: 270970608
Change-Id: Ibc70c8e18081851dfe1c12e129c1da6ab0346a58