Remove more unused CodeIdentity utility methods
This commit continues the SgxIdentity migration by removing
SetSelfCodeIdentity (which is now unused) and moving the CodeIdentity
variant of ParseSgxIdentity to the anonymous namespace.

In addition, some assertions were added to the legacy tests in
`sgx_identity_expectation_matcher_test` to ensure the serialized
SgxIdentities being tested are actually legacy identities.

PiperOrigin-RevId: 271142203
Change-Id: I78301781bc48834366e27c35fce3809f7f71578c