Remove CodeIdentity expectation-matching method
This commit removes the MatchIdentityToExpectation utility method that
operates on CodeIdentity and CodeIdentityExpectation messages, as part
of the ongoing SgxIdentity migration.

It also adds a method to `code_identity_util_test` to assist with
generating minimal SgxIdentityExpectation messages, in an effort to
increase the readability of the unit tests in the file.

PiperOrigin-RevId: 271227346
Change-Id: If955ae1e45e3052dc45de4e010b3e835f6a7c166