Move DeliverSignal down to primitives
DeliverSignal is a sgx specific implementation so move it from trusted
application to primitives. Also change the implementation to pass a
simple string directly instead of MessageReader/MessageWriter, and avoid
going through the asylo_enclave_call interface. Because DeliverSignal is
part of a signal handler, and we should avoid mutex/enclave exit call.

PiperOrigin-RevId: 272490270
Change-Id: Iedcd24650a1d4b7f0b904db7a1abf3b5c624c6ec