Refactor SgxIdentity-testing utility methods
This commit makes a few changes to clean up the methods used to test
the behaviour of `code_identity_util`. Firstly, it adds a default
`is_legacy` parameter to SetExpectation, as this is propagated to the
validity check of the expectation's match spec.

Secondly, it moves the methods that were previously in the internal
namespace of `code_identity_util` into the anonymous namespace, as its
dependencies were rewritten to not require calling them anymore.

Finally, it makes the naming of the methods that generate random and
invalid messages in `code_identity_test_util` more consistent.

PiperOrigin-RevId: 272679457
Change-Id: I5218b5f1d192ec972a41fda5d7e0891b2d02eced