Create sgx_identity_util library
This commit introduces `sgx_identity_util`, which is intended to be a
nicer public-facing API for interacting with SgxIdentity-related
messages. Compared to the already-existing `code_identity_util`, this
library is more opinionated and exposes less of the plumbing that
underlies the rest of the Asylo identity codebase.

Because this library has a dependency on `sgx_identity_cc_proto` which
only has implementation-level visibility, this commit also only sets the
visibility of `sgx_identity_util` to `implementation`, despite it being
intended as a public API. Once the visibility of `sgx_identity_cc_proto`
is updated to `public`, this build target should be updated too, as well
adding it to the Doxyfile so that the header is documented.

PiperOrigin-RevId: 273847346
Change-Id: I566ab17cbc5da2440d0a59c1f48292453470d518