Fix |current_client| in child process after fork
|current_client| is pointing to the current active enclave client, which
is used to identify the correct exit handlers. In case of fork(),
|current_client| is pointing to a deprecated client, while a new enclave
is loaded in the child process and a new client is generated for it.
This commit resets |current_client| to the newly generated client after
a new enclave is loaded in the child process, so that the client can
locate the correct exit handlers.

PiperOrigin-RevId: 273869176
Change-Id: I59f1bbe540e55d6bbdc29d1ab0269dbdf2ac62af