Add a library for ASN.1 schemas
The library consists of composable schema objects. Only schemas that are
expected to be used are supported. Specifically, the ASN.1 schema
library currently only supports schemas that are necessary for parsing
PCK certificate extensions.

The Asn1Sequence() schema currently only supports two arguments. It is
possible to support an arbitrary number of arguments, but the code to do
that is very complex, and only two arguments are needed to parse PCK
certificate extensions.

PiperOrigin-RevId: 275316282
Change-Id: I7b9a89904c2263eba0b54edc57d1cfceeff67c0b