Add select() around trusted dependencies
The select() conditional build feature ensures that the Asylo toolchain
is appropriately in use to build a given target. If a target depends on
a trusted target, it should guard that dependency with a select().

Use of sgx-specific targets that are generic to sgx_sim and sgx_hw
should use target-specific variable definitions to avoid duplication.

PiperOrigin-RevId: 275329315
Change-Id: Ie74f5f5cc7b0fd9f4070d29a152d81f0417a77fb