Move SgxIdentity types into separate proto file
This commit creates a new proto file (`machine_configuration.proto`) to
house subtypes of SgxIdentity and SgxIdentityMatchSpec. This involves
relocating CpuSvn and SgxType from `platform_provisioning.proto` (since
these are more correctly classified as properties of the enclave
identity rather than a provisioned platform), as well as moving the
SgxMachineConfiguration and SgxMachineConfigurationMatchSpec messages
out of `sgx_identity.proto`, removing the `Sgx` prefix in the process
(as they now reside in the `asylo::sgx` namespace).

PiperOrigin-RevId: 275352212
Change-Id: Ib7e4f6637f8cfa4351964b07eb9284ac8ba81491