Migrate client_side_auth_test to sgx_identity_util
This commit migrates `client_side_authorization_test` to depend on
`sgx_identity_util` rather than `code_identity_util`. Because this is
the last remaining instance of code outside of `identity/sgx/` relying
on `code_identity_util`, this makes it safe to lock down the visibility
of that library to the directory-level in the future.

PiperOrigin-RevId: 275545696
Change-Id: I2c4de271d7b44c203408b9b2e210bd6cae68e15f