Replace bridge_siginfo_t, remove bridge_types
Replaces bridge_siginfo_t with kernel type klinux_siginfo_t.
siginfo_t on the untrusted side is explicitly referred to by
klinux_siginfo_t to conform to the function pointer signature
of sigaction inside the enclave and on the host.

Also removes the now unused bridge_types and bridge_functions.

PiperOrigin-RevId: 276357805
Change-Id: I8e75233f502b7eca357530952bd20bea3d83641c