Add asylo_sgx_sim/asylo_sgx_hw backend labels
The new labels will be used when the continuous integration
container runs Bazel 1.1.0, since flag_values could not reference
a label with a workspace name before that release.

The new backend label_flag will be governed by transition rules, soon,
but in the interim, all backend selects will be duplicated to also
respect the old backend configurations based on --define.

PiperOrigin-RevId: 277076502
Change-Id: Ib76285fb9297ca173a9b969021cd06def4c70f2b