Migrate SGX secret sealer to sgx_identity_util
This commit updates methods in `sgx_local_secret_sealer` to call out to
`sgx_identity_util` instead of `sgx_identity_util_internal`, since it
doesn't require any of the internal-only functionality.

In addition, because the ParseKeyGenerationParams call already performs
an sgx::MatchIdentityToExpectation() check against the self identity,
the CPUSVN checks in the tests are redundant and have been removed.
Another test was added to verify the reference identity assigned in
SetDefaultHeader() is equivalent to the current enclave identity (the
original purpose of the aforementioned CPUSVN check).

PiperOrigin-RevId: 277154194
Change-Id: Ife94251a45c87983cebb6d284f1dc174b388261e