Rename ASYLO_SIM to ASYLO_DLOPEN
The "sim" naming is a misnomer, since this backend does not simulate
any particular technology. Instead going forward, we will name backends
by the technology they use. Since the "ASYLO_SIM" backend implements the
primitives interface with a dynamic library wrapper, we are choosing to
call it the "dlopen backend".

PiperOrigin-RevId: 277758515
Change-Id: Ia34c888fa85bf17a02e7e768fb4d12710863708a