Change signal dispatcher mutex to recursive
Signal dispatcher methods are used by signal handler, change it to use
recursive mutex to prevent deadlock when a registered signal arrives
while the thread is holding the lock. This is safe because we are
masking signals to prevent a singal handler from interrupting a thread
while it's modifying it.

PiperOrigin-RevId: 279783700
Change-Id: I3ca33e01b21546658b78f8b91267ab8653322865