Add parser for Intel DCAP packed quote buffer
The Intel DCAP library returns a packed buffer with multiple variable-
length byte arrays nested inside of it. This change adds a new parsing
function which takes the packed buffer and converts it into an array.
The variable-length areas of the quote buffer are copied into vectors.

All fixed-size data structures remain packed and byte-compatible with
the data returned by the QE, so quote verification can still be done
using the unpacked data.

PiperOrigin-RevId: 280015457
Change-Id: I7cdf9cc35d7fb751cc9cd39a9f1b89318431f83b