Make asylo.bzl rules backend-generic
Rules in asylo.bzl take a backends field that specifies
the intended backends to use for defining and signing enclaves. Multiple
backends generate an alias target that selects a backend-specialized
target based on appropriate config_settings. The alias gets tagged with
all the backends' tags, and the specialized targets get tagged with
just the backend's tags. The specialized targets' names are generated
from the base name and the backend label string. Invalid identifier
characters in labels @/: are translated to _.

The dlopen backend is not yet fully supported, so tests that should run
for all backends are currently limited to SGX backends.

Some definitions have a closed-world semantics. The generic unsigned
enclave rule must have all backends' implicit dependencies added as
rule attributes. There is no way for providers to smuggle in implicit
dependencies to make for a truly open-world generic rule implementation.

Furthermore, generic labels for a debug key, signing tool, and signing
configuration are defined in //asylo/bazel in a closed-world way. For
unknown backends, the default key and config is an empty file. The
default sign tool is the "true" program (returns a 0 exit code).

PiperOrigin-RevId: 280046519
Change-Id: I891ca1c0cfde14c40c77d037d7db5d6289f2d8fd