Remove unnecessary restriction on sgx-sim PCE
The code that simulates launch token fetching had an additional
restriction for enclaves with the PROVISION_KEY attribute. This
effectively made it impossible for anyone but Intel to create and
sign a simulated PCE.

Drop this restriction so that others can build and use the PCE with
the simulated SGX instructions. This allows testing of the provisioning
and certification flow on systems without SGX hardware.

PiperOrigin-RevId: 280062682
Change-Id: I60c3c647bb4e16a41f15c53aeb77a5d9af30ac2f