Move copy_from_host into its own bzl file
The copy_from_host rule will be deprecated in favor of Bazel
configuration transitions. The feature is still experimental and does
not support transitions from external rules, so it is initially
disabled.

To smoothly deprecate the rule, we conditionally choose an old rule
or a new rule based on the package of the requesting rule. This commit
creates a file which will keep the "old" rules.

PiperOrigin-RevId: 280275403
Change-Id: I2550f08c01756e017beecba4ca229d2ea7d8f886