Rename kSgxRemoteAssertionAuthority constant
This commit adds "Age" to the name of `kSgxRemoteAssertionAuthority`
and the related helper method used to set this description. The
reason for this change is that both the Assertion Generator Enclave
(AGE) and the Intel QE will have their own authorities, and there is
a need to differentiate between the two.

PiperOrigin-RevId: 280689227
Change-Id: Ib04ab83b7e89b11b0527921db91ba591e0776c3c