Hold entry block in fork until fork() on host
Currently we are holding enclave entry block while taking a snapshot of
the enclave during fork(). They are unblocked after the snapshot is
taken. However, at that point fork() on the host is not invoked yet,
which may cause enclave snapshot and host memory to be inconsistent in
the child process. This change extends the block period to when the
parent process starts to transfer key. At this point fork() on the host
is invoked as well.
Also, before we are using an RAII wrapper during snapshot to unblock the
entries while taking the snapshot. This introduces security issues as
this wrapper is copied to the snapshot and may unblock entries in the
child process while not intended to.
This also puts the condition of other threads in a controllable state so
that they won't be holding enclave entry/exit locks on the untrusted
side while fork() on the host is invoked.

PiperOrigin-RevId: 280711197
Change-Id: I805ae8b90fa602e15843b85177d7170cf0cc7a12