Allow generating AGE key without TargetInfo
This commit updates the GenerateKeyAndCsr method of the assertion
generator enclave to have the ability to generate a key (and thus be
initialized) even when no PCE TargetInfo was passed in as an input.

This change makes integration testing flows between the AGE and other
components simpler by removing the requirement of passing a valid
TargetInfo proto before being able to interface with the AGE at all.

PiperOrigin-RevId: 281120419
Change-Id: I5614bd33cb5a5fb5a1bd3957c0d19de5e6d2d067