Implement SgxAgeRemoteAssertionGenerator library
This commit implements SgxAgeRemoteAssertionGenerator, a library that
wraps SgxRemoteAssertionGeneratorClient. It also creates an
SgxAgeRemoteAssertionAuthorityConfig proto, used to specify the
configuration for the SGX AGE remote assertion authority.

While it would be ideal for the gRPC channel/client to be established
only a single time within the generator's Initialize method, when
working on this commit, there was odd and difficult-to-replicate
behaviour when it came to long-running channels connected to the AGE,
resulting in fairly flaky tests. While the underlying issue was fixed
before merging this commit, being weary of channels suddenly dying,
it seems safer to just create the channel right when it is needed.

Despite the gRPC documentation stating that establishing a channel is
a "heavyweight" operation, preliminary benchmarking suggests that
Generate() still runs in under 50ms, so the decision was made to
check in this implementation as-is, and changes can be made in the
future if performance needs to be tuned.

The approach to testing the functionality of the generator is to spin
up one instance of the AGE in the SetUpTestSuite part of the tests in
an enclave, and call out to a "helper enclave" that is responsible
for running the generator. Without running both parts of the program
in an enclave, local attestation will fail, and the channel between
the AGE and the generator will not be able to be established.

PiperOrigin-RevId: 281813431
Change-Id: Id532f932b29a78adf6ece2783851fbf8a96eba7a