Rename generate_sigstruct build rule
The new name, generate_enclave_signing_material, is closer to
its intended function. The SIGSTRUCT, enclave_css_t type, contains
this signing material, but also a signature of it, the signing key's
public key, and some derived information from the public key. Therefore
calling the output of this rule a sigstruct is a misnomer.

Resolves #51.

PiperOrigin-RevId: 281841345
Change-Id: Ie71b45207bcf9484b4f8cdd6619c5efbbe7aeb1d