Load AGE test enclave once in SetUpTestSuite
With the current sgx_age_remote_assertion_generator tests, the test
enclave is loaded once per test fixture. When testing on actual SGX
hardware, it takes approximately 5 seconds just to launch the test
enclave; since this happens once per test fixture, it ends up taking
around 2 minutes just to get through the entire suite.

This commit changes the testing approach to load the test enclave a
single time in the SetUpTestSuite method (similar to how the AGE
enclave itself is only loaded a single time), and adds an additional
method to the test enclave to clear the generator instance that it
holds, which is invoked in each of the test fixtures' SetUp. With
this change, the tests only take on the order of 10 seconds to run
on SGX hardware, which is a significant improvement.

PiperOrigin-RevId: 282029941
Change-Id: I87bc3d2d897eff0e8cc6f3268204f7e8fda4a708