Add prefix to map_file and unmap_file
These functions are defined in both linux_sgx and
sgx_data_center_attestation_primitives. When statically linking from
both sources, the wrong functions were sometimes picked up, depending
on linker order.

What's worse, the version of map_file in linux_sgx is incompatible
with map_file from sgx_data_center_attestation_primitives. The result
was smashed stacks, as the linux_sgx version of the function writes to a
64-bit output variable, but sgx_data_center_attestation_primitives only
passes a pointer to 32-bits of storage.

Prefix the functions in sgx_data_center_attestation_primitives with
"dcap_internal_" so that we do not get namespace collisions.

PiperOrigin-RevId: 282579425
Change-Id: I673ec2a9fd328d3f295253db0c38afba81c8c4cd