Update the aad generator interface
The AdditionalAuthenticatedDataGenerator factory functions used
to return a StatusOr. However, they didn't need to, as the data used
in construction is fixed. This change updates the factories so they
will always return a unique_ptr, instead of returning a unique_ptr
wrapped in StatusOr.

Make the member functions const, because they change no object
state.

Change Generate so it accepts a ByteContainerView, which is implicitly
constructable from various byte containers.

Change Generate to return UnsafeBytes, which makes use of the generator
more natural, as the output is being copied into byte arrays.

PiperOrigin-RevId: 282642939
Change-Id: I2b6f7b08e93100b071ed6526fa57249342648923