Make asylo_package stricter
The workspace_name mapping issue is still unresolved in Bazel, so
@com_google_asylo// and // are different when within Asylo. Checking
"asylo" in package_name unfortunately is too lenient and breaks for
external users that have an "asylo" subdirectory. As long as a user
does not have "//asylo" as part of their package, this fix should work.
There is no clean semantically correct solution until
--incompatible_remap_main_repo is flipped.

The generic enclave rules use a couple default labels that needed
public visibility, so those are changed here too.

Resolves #52

PiperOrigin-RevId: 283666824
Change-Id: Ie793b9c090d83c15c1fd571bac9bef3f38d1274e