Add fork consistency test
Add a test that invokes fork() while other threads keep updating random
memory in random buffers, and check in the child to confirm that the
values in the buffers are consistent. This test is to check whether the
child memory is a consistent snapshot of the parent at a certain point.

PiperOrigin-RevId: 284294335
Change-Id: I360075f5d4a1bab9b0b48fd616e8a41f0aa45d72