Move enclave_exit to primitives
Currently enclave_exit goes through enclave UntrustedCall to exit the
enclave and calls exit on the host. This is not an async-signal-safe
process when used with fork(2), because it obtains locks, allocate and
free memories during the process, which could cause deadlock if other
threads are holding the locks or allocating/freeing memories while
fork(2) is invoked. On a linux host _exit(2) is supposed to be safe to
call after fork(2). Therefore move enclave_exit to primitives to exit
the enclave directly without allocating/locking.

PiperOrigin-RevId: 285028432
Change-Id: I00b5dbbd0aa37ed1a2b652f56d7098847c2f5d8f