Give debug_sign_enclave open world semantics
The previous design required a select() over all supported backends for
a default key, config, and sign_tool. This change allows for each
backend to specify their own defaults as arguments via a backend label
struct. A backend label struct is an existing concept that was created
to make backend_tools.all_backends easy to use for arbitrary backends.

PiperOrigin-RevId: 285100609
Change-Id: I8779ab4afb4f4439d31e92c34973c8107057cfb1